% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowSorted.Blood.WGBS.BLUEPRINT.R
\docType{data}
\name{FlowSorted.Blood.WGBS.BLUEPRINT}
\alias{FlowSorted.Blood.WGBS.BLUEPRINT}
\title{Whole genome bisulfite-sequencing data on sorted 
blood cell populations from BLUEPRINT}
\format{A BSseq object with 44 WGBS samples and 
approximately 29 million CpGs (or 29039352 CpGs
to be exact).}
\usage{
FlowSorted.Blood.WGBS.BLUEPRINT()
}
\description{
This package measures the DNA methylation of
sorted cell types from whole blood (venous and cord) 
from the BLUEPRINT consortia using the whole genome 
bisulfite-sequencing platform (WGBS). BigWig files
were downloaded and converted to HDF5 files. Data 
are available on ExperimentHub as a data package.

The HDF5 files were created using 
the files: 
     /inst/scripts/bp-01-create-data-object.R
     /inst/scripts/bp-02-download-data.sh
     /inst/scripts/bp-03-create-hdf5-bsseq-object.R
and the object downloaded from ExperimentHub.
}
\examples{
library(ExperimentHub)
library(FlowSorted.Blood.WGBS.BLUEPRINT)
fs_wgbs <- FlowSorted.Blood.WGBS.BLUEPRINT()
dim(fs_wgbs)

}
